@@ -14,53 +14,59 @@
 # You should have received a copy of the GNU General Public License along
 # with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from typing import Optional, cast
+"""Gadget plugin."""
+
+from typing import Any, cast
+
 from craft_parts import plugins
+from typing_extensions import Self
 
 from imagecraft.utils import craft_base_to_ubuntu_series
 
 
 class GadgetPluginProperties(plugins.PluginProperties):
-    gadget_target: Optional[str] = None
+    """Supported attributes for the 'gadget' plugin."""
+
+    gadget_target: str | None = None
 
     @classmethod
-    def unmarshal(cls, data):
-        plugin_data = plugins.base.extract_plugin_properties(
-            data, plugin_name="gadget"
-        )
+    def unmarshal(cls, data: dict[str, Any]) -> Self:
+        """unmarshal."""
+        plugin_data = plugins.base.extract_plugin_properties(data, plugin_name="gadget")
         return cls(**plugin_data)
 
 
 class GadgetPlugin(plugins.Plugin):
+    """Builds the gadget containing bootloader configuration."""
+
     properties_class = GadgetPluginProperties
 
-    def get_build_snaps(self):
-        return {}
+    def get_build_snaps(self) -> set[str]:
+        """Return a set of required snaps to install in the build environment."""
+        return set()
 
-    def get_build_packages(self):
-        # This list here should include all the base packages that are needed
-        # to build the gadget. More specific packages to the actual gadget
-        # build should be added manually by users.
+    def get_build_packages(self) -> set[str]:
+        """Return a set a packages needed to build the gadget.
+
+        More specific packages to the actual gadget build should be added manually by users.
+        """
         return {"make"}
 
-    def get_build_environment(self):
+    def get_build_environment(self) -> dict[str, Any]:
+        """Return a dictionary with the environment to use in the build step."""
         gadget_arch = self._part_info.target_arch
-        gadget_series = craft_base_to_ubuntu_series(
-            self._part_info.project_info.base)
+        gadget_series = craft_base_to_ubuntu_series(self._part_info.project_info.base)
         return {"ARCH": gadget_arch, "SERIES": gadget_series}
 
-    def get_build_commands(self):
+    def get_build_commands(self) -> list[str]:
+        """Return a list of commands to run during the build step."""
         options = cast(GadgetPluginProperties, self._options)
 
         gadget_target = options.gadget_target
         if gadget_target is None:
             gadget_target = ""
 
-        gadget_cmd = [
+        return [
             f"make {gadget_target}",
             "cp -a $CRAFT_PART_BUILD/install/* $CRAFT_PART_INSTALL/",
-            # Temporary before we change ubuntu-image
-            "cp $CRAFT_PART_INSTALL/meta/gadget.yaml $CRAFT_PART_INSTALL/",
         ]
-
-        return gadget_cmd
