@@ -6,9 +6,35 @@
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
-project = "imagecraft"
+project = "Imagecraft"
 copyright = "2023, Canonical"
-author = "Canonical"
+author = "Canonical Ltd."
+
+# Update with the favicon for your product
+# html_favicon = "_static/favicon.png"
+
+html_context = {
+    # Change to the link to your product website (without "https://")
+    "product_page": "github.com/canonical/imagecraft",
+    # Add your product tag to ".sphinx/_static" and change the path
+    # here (start with "_static"), default is the circle of friends
+    "product_tag": "_static/tag.png",
+    # Change to the discourse instance you want to be able to link to
+    # (use an empty value if you don't want to link)
+    # "discourse": "https://discourse.ubuntu.com/c/rocks/117",
+    # Change to the GitHub info for your project
+    "github_url": "https://github.com/canonical/imagecraft",
+    # Change to the branch for this version of the documentation
+    "github_version": "main",
+    # Change to the folder that contains the documentation (usually "/" or "/docs/")
+    "github_folder": "/docs/",
+    # Change to an empty value if your GitHub repo doesn't have issues enabled
+    "github_issues": "enabled",
+}
+
+
+if "discourse" in html_context:
+    html_context["discourse_prefix"] = html_context["discourse"] + "/t/"
 
 # region General configuration
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
@@ -21,13 +47,25 @@ extensions = [
     "sphinx_design",
     "sphinx_copybutton",
     "sphinx-pydantic",
-    "sphinx_toolbox",
-    "sphinx_toolbox.more_autodoc",
+    # "sphinx_toolbox.more_autodoc",
     "sphinx.ext.autodoc",  # Must be loaded after more_autodoc
 ]
 
+
+# List of patterns, relative to source directory, that match files and
+# directories to ignore when looking for source files.
+# This pattern also affects html_static_path and html_extra_path.
+exclude_patterns = [
+    "_build",
+    "Thumbs.db",
+    ".DS_Store",
+    "env",
+    "sphinx-starter-pack",
+    "common",
+]
+
 templates_path = ["_templates"]
-exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
+
 
 show_authors = False
 
@@ -35,10 +73,81 @@ show_authors = False
 # region Options for HTML output
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
+# The theme to use for HTML and HTML Help pages.  See the documentation for
+# a list of builtin themes.
+#
 html_theme = "furo"
+html_last_updated_fmt = ""
+html_permalinks_icon = "¶"
+html_theme_options = {
+    "light_css_variables": {
+        "color-sidebar-background-border": "none",
+        "font-stack": "Ubuntu, -apple-system, Segoe UI, Roboto, Oxygen, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif",
+        "font-stack--monospace": "Ubuntu Mono, Consolas, Monaco, Courier, monospace",
+        "color-foreground-primary": "#111",
+        "color-foreground-secondary": "var(--color-foreground-primary)",
+        "color-foreground-muted": "#333",
+        "color-background-secondary": "#FFF",
+        "color-background-hover": "#f2f2f2",
+        "color-brand-primary": "#111",
+        "color-brand-content": "#06C",
+        "color-api-background": "#cdcdcd",
+        "color-inline-code-background": "rgba(0,0,0,.03)",
+        "color-sidebar-link-text": "#111",
+        "color-sidebar-item-background--current": "#ebebeb",
+        "color-sidebar-item-background--hover": "#f2f2f2",
+        "toc-font-size": "var(--font-size--small)",
+        "color-admonition-title-background--note": "var(--color-background-primary)",
+        "color-admonition-title-background--tip": "var(--color-background-primary)",
+        "color-admonition-title-background--important": "var(--color-background-primary)",
+        "color-admonition-title-background--caution": "var(--color-background-primary)",
+        "color-admonition-title--note": "#24598F",
+        "color-admonition-title--tip": "#24598F",
+        "color-admonition-title--important": "#C7162B",
+        "color-admonition-title--caution": "#F99B11",
+        "color-highlighted-background": "#EbEbEb",
+        "color-link-underline": "var(--color-background-primary)",
+        "color-link-underline--hover": "var(--color-background-primary)",
+        "color-version-popup": "#772953",
+    },
+    "dark_css_variables": {
+        "color-foreground-secondary": "var(--color-foreground-primary)",
+        "color-foreground-muted": "#CDCDCD",
+        "color-background-secondary": "var(--color-background-primary)",
+        "color-background-hover": "#666",
+        "color-brand-primary": "#fff",
+        "color-brand-content": "#06C",
+        "color-sidebar-link-text": "#f7f7f7",
+        "color-sidebar-item-background--current": "#666",
+        "color-sidebar-item-background--hover": "#333",
+        "color-admonition-background": "transparent",
+        "color-admonition-title-background--note": "var(--color-background-primary)",
+        "color-admonition-title-background--tip": "var(--color-background-primary)",
+        "color-admonition-title-background--important": "var(--color-background-primary)",
+        "color-admonition-title-background--caution": "var(--color-background-primary)",
+        "color-admonition-title--note": "#24598F",
+        "color-admonition-title--tip": "#24598F",
+        "color-admonition-title--important": "#C7162B",
+        "color-admonition-title--caution": "#F99B11",
+        "color-highlighted-background": "#666",
+        "color-link-underline": "var(--color-background-primary)",
+        "color-link-underline--hover": "var(--color-background-primary)",
+        "color-version-popup": "#F29879",
+    },
+}
+
+# Add any paths that contain custom static files (such as style sheets) here,
+# relative to this directory. They are copied after the builtin static files,
+# so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ["_static"]
+
+# These paths are either relative to html_static_path
+# or fully qualified paths (eg. https://...)
 html_css_files = [
     "css/custom.css",
+    "github_issue_links.css",
+    "custom.css",
+    "header.css",
 ]
 
 # endregion
@@ -54,9 +163,12 @@ intersphinx_mapping = {
 set_type_checking_flag = True
 typehints_fully_qualified = False
 always_document_param_types = True
+typehints_document_rtype = True
+linkcheck_anchors_ignore = ["slice-definitions"]
+
+# Enable support for google-style instance attributes.
+napoleon_use_ivar = True
 
 # Github config
 github_username = "canonical"
 github_repository = "imagecraft"
-
-# endregion
