@@ -14,6 +14,8 @@
 import types
 
 import pytest
+from craft_application.models.constraints import ProjectName, SummaryStr
+from imagecraft import plugins
 
 
 @pytest.fixture()
@@ -31,6 +33,62 @@ def project_main_module() -> types.ModuleType:
         main_module = imagecraft
     except ImportError:
         pytest.fail(
-            "Failed to import the project's main module: check if it needs updating"
+            "Failed to import the project's main module: check if it needs updating",
         )
     return main_module
+
+
+@pytest.fixture(autouse=True, scope="session")
+def _setup_parts():
+    plugins.setup_plugins()
+
+
+@pytest.fixture()
+def new_dir(monkeypatch, tmpdir):
+    """Change to a new temporary directory."""
+    monkeypatch.chdir(tmpdir)
+    return tmpdir
+
+
+@pytest.fixture()
+def extra_project_params():
+    """Configuration fixture for the Project used by the default services."""
+    return {}
+
+
+@pytest.fixture()
+def default_project(extra_project_params):
+    from craft_application.models.constraints import VersionStr
+    from imagecraft.models.project import Platform, Project
+
+    parts = extra_project_params.pop("parts", {})
+
+    return Project(
+        name=ProjectName("default"),
+        version=VersionStr("1.0"),
+        summary=SummaryStr("default project"),
+        description="default project",
+        base="ubuntu@22.04",
+        series="jammy",
+        parts=parts,
+        platforms={"amd64": Platform(build_for=["amd64"], build_on=["amd64"])},
+        **extra_project_params,
+    )
+
+
+@pytest.fixture()
+def default_factory(default_project):
+    from imagecraft.application import APP_METADATA
+    from imagecraft.services import ImagecraftServiceFactory
+
+    return ImagecraftServiceFactory(
+        app=APP_METADATA,
+        project=default_project,
+    )
+
+
+@pytest.fixture()
+def default_application(default_factory):
+    from imagecraft.application import APP_METADATA, Imagecraft
+
+    return Imagecraft(APP_METADATA, default_factory)
